main: move java classes to Java/ folder
