+++ Java/BinaryNumberOneToN.java
@@ -6,7 +6,7 @@ For example, for n = 16, the binary numbers are:
 [1 10 11 100 101 110 111 1000 1001 1010 1011 1100 1101 1110 1111 10000]
  */
 
-package Queue;
+package Java;
 
 import java.util.Arrays;
 import java.util.LinkedList;
