+++ Java/armstrongNumber.java
@@ -1,3 +1,5 @@
+package Java;
+
 import java.util.Scanner;
 public class armstrongNumber
 {
