+++ Java/friendlypair.java
@@ -1,3 +1,5 @@
+package Java;
+
 public class Main
  {
    public static void main (String[]args)
