+++ Java/rhombus_start_pattern.java
@@ -1,3 +1,5 @@
+package Java;
+
 import java.util.Scanner;
 public class Pattern1 {
